+++ tests/run_examples.sh
@@ -89,6 +89,7 @@ examples=(
     path_builtins_test.bas
     loop_syntax_test.bas
     else_if_test.bas
+    datetime_zone_test.bas
     count_builtin_test.bas
     string_concat_test.bas
     parse_test.gb